Move format specific file classes to their appropriate parts package
