+++ src/photon/application/dialogs/SaveDialog.java
@@ -28,8 +28,8 @@ import com.intellij.uiDesigner.core.GridConstraints;
 import com.intellij.uiDesigner.core.GridLayoutManager;
 import com.intellij.uiDesigner.core.Spacer;
 import photon.application.MainForm;
-import photon.file.PhotonFile;
-import photon.file.Sl1File;
+import photon.file.parts.photon.PhotonFile;
+import photon.file.parts.sl1.Sl1File;
 import photon.file.SlicedFile;
 import photon.file.parts.EFileType;
 import photon.file.parts.IFileHeader;
